move: User관련 enum 폴더 생성 (#19)
