+++ src/main/java/zighang2/zighang/web/domain/user/enums/Transport.java
@@ -1,0 +1,6 @@
+package zighang2.zighang.web.domain.user.enums;
+
+public enum Transport {
+    CAR,
+    TRANSIT
+}
